Rename *_stores
