+++ terracotta/drivers/geotiff_raster_store.py
